--- src/hooks/useCartItems.js
@@ -1,9 +1,0 @@
-import { useEffect } from "react";
-
-export const useCartItems = (cartItems, setCartCount, setTotal) => {
-  useEffect(() => {
-    setCartCount(cartItems.length)
-    setTotal(+(cartItems.reduce((curr, { price }) => curr + price, 0)).toFixed(1))
-    localStorage.setItem('cartItems', JSON.stringify(cartItems));
-  }, [cartItems])
-}
