--- src/hooks/useFavorites.js
@@ -1,7 +1,0 @@
-import { useEffect } from "react";
-
-export const useFavorites = (favorites) => {
-  useEffect(() => {
-    localStorage.setItem('favorites', JSON.stringify(favorites));
-  }, [favorites])
-}
