--- src/hooks/useOrders.js
@@ -1,7 +1,0 @@
-import { useEffect } from "react"
-
-export const useOrders = orders => {
-  useEffect(() => {
-    localStorage.setItem('orders', JSON.stringify(orders));
-  }, [orders])
-}
